hrms_backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg2-binary
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
